epgcache: fix build when using CI
This commit fixes the following error:
dvb/epgcache.cpp:2208:46: error: cannot convert ‘int*’ to ‘Py_ssize_t*’ {aka ‘long int*’}

+++ lib/dvb/epgcache.cpp
@@ -2190,7 +2190,7 @@ PyObject *eEPGCache::search(ePyObject arg)
 	int eventid = -1;
 	const char *argstring=0;
 	const char *refstr=0;
-	int argcount=0;
+	ssize_t argcount=0;
 	int querytype=-1;
 	bool needServiceEvent=false;
 	int maxmatches=0;
